Extract game commands to upper classes
