+++ HTW/src/htw/game/RestCommand.java
@@ -1,0 +1,10 @@
+package htw.game;
+
+class RestCommand extends GameCommand {
+  public RestCommand(HuntTheWumpusGame game) {
+    super(game);
+  }
+
+  public void processCommand() {
+  }
+}
